mint shares when the we increase the split

+++ src/EtherFiAdmin.sol
@@ -168,7 +168,8 @@ contract EtherFiAdmin is Initializable, OwnableUpgradeable, UUPSUpgradeable {
     //protocol owns the eth that was distributed to NO and treasury in eigenpods and etherfinodes 
     function _handleProtocolFees(IEtherFiOracle.OracleReport calldata _report) internal { 
         require(_report.protocolFees >= 0, "EtherFiAdmin: protocol fees can't be negative");
-        liquidityPool.payProtocolFees(_report.protocolFees);
+        //verify that protocol fee is never more 10% fees
+        liquidityPool.payProtocolFees(uint128(_report.protocolFees));
     }
 
     function _handleAccruedRewards(IEtherFiOracle.OracleReport calldata _report) internal {
